+++ json_serializable/test/test_sources/test_sources.dart
@@ -16,6 +16,7 @@ class ConfigurationImplicitDefaults {
   createFactory: true,
   createToJson: true,
   createFieldMap: false,
+  createPerFieldToJson: false,
   disallowUnrecognizedKeys: false,
   explicitToJson: false,
   fieldRename: FieldRename.none,